+++ addon/plugins/article-structure-plugin/commands/insert-structure.ts
@@ -2,19 +2,13 @@ import {
   Command,
   Fragment,
   NodeSelection,
-  NodeType,
-  PNode,
-  Schema,
-  Selection,
   TextSelection,
 } from '@lblod/ember-rdfa-editor';
 import recalculateStructureNumbers from './recalculate-structure-numbers';
 import { StructureSpec } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/article-structure-plugin';
 import wrapStructureContent from './wrap-structure-content';
 import IntlService from 'ember-intl/services/intl';
-import { findNodes } from '@lblod/ember-rdfa-editor/utils/position-utils';
-import { containsOnlyPlaceholder } from '../utils/structure';
-import { findParentNodeOfType } from '@curvenote/prosemirror-utils';
+import { findInsertionRange } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/_private/find-insertion-range';
 
 const insertStructure = (
   structureSpec: StructureSpec,
@@ -28,7 +22,7 @@ const insertStructure = (
     }
     const insertionRange = findInsertionRange({
       doc,
-      selection,
+      $from: selection.$from,
       nodeType: schema.nodes[structureSpec.name],
       schema,
       limitTo: structureSpec.limitTo,
@@ -64,76 +58,5 @@ const insertStructure = (
     return true;
   };
 };
-
-function findInsertionRange(args: {
-  doc: PNode;
-  selection: Selection;
-  nodeType: NodeType;
-  schema: Schema;
-  limitTo?: string;
-}) {
-  const { doc, selection, nodeType, schema, limitTo } = args;
-  const { $from } = selection;
-  for (let currentDepth = $from.depth; currentDepth >= 0; currentDepth--) {
-    const currentAncestor = $from.node(currentDepth);
-    const index = $from.index(currentDepth);
-    if (currentAncestor.canReplaceWith(index, index, nodeType)) {
-      if (containsOnlyPlaceholder(schema, currentAncestor)) {
-        return { from: $from.start(currentDepth), to: $from.end(currentDepth) };
-      } else {
-        const insertPos = $from.after(currentDepth + 1);
-        return { from: insertPos, to: insertPos };
-      }
-    }
-  }
-  const limitContainer = limitTo
-    ? findParentNodeOfType(schema.nodes[limitTo])(selection)
-    : null;
-
-  const limitContainerRange = limitContainer
-    ? {
-        from: limitContainer.pos,
-        to: limitContainer.pos + limitContainer.node.nodeSize,
-      }
-    : { from: 0, to: doc.nodeSize };
-  const filterFunction = ({ from, to }: { from: number; to: number }) => {
-    if (from >= limitContainerRange.from && to <= limitContainerRange.to) {
-      const node = doc.nodeAt(from);
-      if (node) {
-        if (node.canReplaceWith(node.childCount, node.childCount, nodeType)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  };
-  const nextContainerRange =
-    findNodes({
-      doc,
-      start: selection.from,
-      visitParentUpwards: true,
-      reverse: false,
-      filter: filterFunction,
-    }).next().value ??
-    findNodes({
-      doc,
-      start: selection.from,
-      visitParentUpwards: true,
-      reverse: true,
-      filter: filterFunction,
-    }).next().value;
-  if (nextContainerRange) {
-    const { from, to } = nextContainerRange;
-    const containerNode = doc.nodeAt(from);
-    if (containerNode) {
-      if (containsOnlyPlaceholder(schema, containerNode)) {
-        return { from: from + 1, to: to - 1 };
-      } else {
-        return { from: to - 1, to: to - 1 };
-      }
-    }
-  }
-  return null;
-}
 
 export default insertStructure;